signup/requirements.txt
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.8.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==1.9
defusedxml==0.5.0
dicttoxml==1.7.4
diff-match-patch==20121119
Django==2.0.2
django-ckeditor==5.7.1
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-crontab==0.7.1
django-debug-toolbar==2.0
django-filter==1.1.0
django-formtools==2.1
django-guardian==1.4.9
django-import-export==1.2.0
django-js-asset==1.2.2
django-rest-swagger==2.2.0
django-reversion==2.0.13
django-simple-history==2.7.3
django-simpleui==2.6
django-validator==0.2.7
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
djangorestframework-jwt-proxy==0.1
drf-extensions==0.3.1
drf-jwt-wechat==1.11.9
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
idna==2.6
importlib-metadata==0.20
inflection==0.3.1
itypes==1.1.0
jdcal==1.3
Jinja2==2.10
jsonschema==3.0.2
jwt==0.5.3
Markdown==2.6.11
MarkupSafe==1.0
more-itertools==7.2.0
oauthlib==2.0.7
odfpy==1.3.6
openapi-codec==1.3.2
openpyxl==2.5.2
optionaldict==0.1.1
packaging==19.1
Pillow==5.1.0
pluggy==0.12.0
py==1.8.0
pycparser==2.18
pycryptodome==3.6.1
PyJWT==1.7.1
PyMySQL==0.8.0
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.1.2
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.3
PyYAML==3.12
raven==6.10.0
requests==2.18.4
requests-oauthlib==0.8.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
simple-history==1.1
simplejson==3.16.0
six==1.11.0
snipeit==1.1
social-auth-app-django==2.1.0
social-auth-core==1.7.0
soupsieve==1.9.3
sqlparse==0.3.0
swagger-spec-validator==2.4.3
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
wechatpy==1.8.3
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0
xmltodict==0.12.0
zipp==0.6.0
